requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.19.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.0.0
python-telegram-bot==21.6
requests==2.32.5
sniffio==1.3.1
urllib3==2.5.0
WooCommerce==3.0.0
yarl==1.20.1
pytz==2025.2